craw_worker/requirements.txt
python-socketio[client]
seleniumbase
beautifulsoup4
lxml
python-dotenv
